User Module

+++ app/Repositories/CarRepository.php
@@ -36,16 +36,42 @@ class CarRepository extends BaseRepository
 
 			$this->setModel($car);
 
-			$this->setSuccess('Successfully save new car.');
+			$this->setSuccess('Successfully save car data.');
 		} catch (QueryException $qe) {
-			$this->setError('Failed to save new car.');
+			$this->setError(
+				'Failed to save car data.', 
+				$qe->getMessage()
+			);
 		}
+
+		return $this->getModel();
+	}
+
+	public function setCarImage($imageFile)
+	{
+		try {
+			$car = $this->getModel();
+			$car->car_image = $imageFile;
+			$car->save();
+
+			$this->setModel($car);
+
+			$this->setSuccess('Successfully set car image.');
+		} catch (QueryException $qe) {
+			$this->setError('Failed to set car image', $qe->getMessage());
+		}
+
+		return $this->getModel();
 	}
 
 	public function delete(bool $forceDelete = false)
 	{
 		try {
 			$car = $this->getModel();
+
+			if ($car->status != 'free') 
+				return $this->setForbidden('Failed to delete car, car in use.');
+
 			$forceDelete ?
 				$car->forceDelete() :
 				$car->delete();
@@ -54,7 +80,12 @@ class CarRepository extends BaseRepository
 
 			$this->setSuccess('Successfully delete car.');
 		} catch (QueryException $qe) {
-			$this->setError('Failed to delete car.');
+			$this->setError(
+				'Failed to delete car.', 
+				$qe->getMessage()
+			);
 		}
+
+		return $this->returnResponse();
 	}
 }
